test: split test suite into sections
Also excludes tests and snapshots from packaging.
